[device_info_plus] rename Android class & package (#56)
