+++ packages/device_info_plus/android/build.gradle
@@ -1,4 +1,4 @@
-group 'io.flutter.plugins.deviceinfo'
+group 'dev.fluttercommunity.plus.device_info'
 version '1.0-SNAPSHOT'
 
 buildscript {
