+++ packages/device_info_plus/android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="io.flutter.plugins.deviceinfo">
+  package="dev.fluttercommunity.plus.device_info">
 </manifest>
