+++ packages/device_info_plus/android/src/main/java/dev/fluttercommunity/plus/device_info/DeviceInfoPlusPlugin.java
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package io.flutter.plugins.deviceinfo;
+package dev.fluttercommunity.plus.device_info;
 
 import android.content.Context;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
@@ -10,14 +10,14 @@ import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
 
-/** DeviceInfoPlugin */
-public class DeviceInfoPlugin implements FlutterPlugin {
+/** DeviceInfoPlusPlugin */
+public class DeviceInfoPlusPlugin implements FlutterPlugin {
 
   MethodChannel channel;
 
   /** Plugin registration. */
   public static void registerWith(Registrar registrar) {
-    DeviceInfoPlugin plugin = new DeviceInfoPlugin();
+    DeviceInfoPlusPlugin plugin = new DeviceInfoPlusPlugin();
     plugin.setupMethodChannel(registrar.messenger(), registrar.context());
   }
