+++ packages/device_info_plus/pubspec.yaml
@@ -9,8 +9,8 @@ flutter:
   plugin:
     platforms:
       android:
-        package: io.flutter.plugins.deviceinfo
-        pluginClass: DeviceInfoPlugin
+        package: dev.fluttercommunity.plus.device_info
+        pluginClass: DeviceInfoPlusPlugin
       ios:
         pluginClass: FLTDeviceInfoPlugin
       linux:
